Add alias for native-comp-deferred deny list
On Emacs versions > 28 the name of the deny list has changed, but
currently `nix-doom-emacs` is pinned to an older version of Doom Emacs
that doesn't have a fix for this.

This change just adds a `defvaralias` in places where it doesn't
exist using feature testing (instead of checking the Emacs version).

This solution was originally developed by @librephoenix.